requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
msal==1.24.0
azure-identity==1.15.0
msgraph-sdk==1.0.0